230102_order 생성_ebko

+++ src/main/java/com/ezen/makingbaking/entity/OrderItemId.java
@@ -6,6 +6,6 @@ import lombok.Data;
 
 @Data
 public class OrderItemId implements Serializable {
-	private int orderNo;
+	private long orderNo;
 	private int itemNo;
 }
